pyproject.toml
[tool.poetry]
name = "ez-sendgrid"
version = "1.0.0"
description = "Manage your SendGrid templates easy."
authors = ["Alex Khaerov <i@hayorov.ru>"]

[tool.poetry.dependencies]
python = "^3.5"
sendgrid = "^6.0.5"
pyyaml = "^5.1.2"
python-decouple = "^3.1"
fire = "^0.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
ez-sendgrid = 'ez_sendgrid.cmd:main'
